fix(M3,M6): 修复丙缺陷报告全部8个缺陷，回归12/12通过
- M3-01 BookServiceImpl三个分页方法pageSize<1容错为默认值（修除零/负数LIMIT的500）
- M3-02 ClientBookServlet翻页URL的nameorauthor参数URLEncoder编码
- M3-03 pageByPrice对min>max自动交换区间，index.jsp显示可见提示
- M3-04 价格区间参数改WebUtils.parseDouble按小数解析，DAO/Service链路改double
- M3-05 榜单SQL LIMIT 1,50改为LIMIT 50（不再跳过销量第一名）
- M6-01 BookServlet.add/update增加validateBook服务端校验（空名/负价/非法价格拒绝），book_edit.jsp显示提示
- M6-02 delete改parseInt容错+按影响行数提示，book_manager.jsp显示结果提示
- M6-03 ManagerUserServlet.add新增existsUsername预检，user_edit.jsp显示提示

回归：tests/run-tests.bat 全部12条用例通过（此前5 OK/7 NG）
验证：pageSize=0→200；min>max→提示+交换；56.5精确命中《三体》；翻页链接已编码

+++ 网上书店项目/src/com/yj/dao/BookDao.java
@@ -22,9 +22,9 @@ public interface BookDao {
 
     List<Book> queryForPageItems(int begin, int pageSize);
 
-    Integer queryForPageTotalCountByPrice(int min, int max);
+    Integer queryForPageTotalCountByPrice(double min, double max);
 
-    List<Book> queryForPageItemsByPrice(int begin, int pageSize, int min, int max);
+    List<Book> queryForPageItemsByPrice(int begin, int pageSize, double min, double max);
 
     Integer queryForPageTotalCountByNameOrAuthor(String nameorauthor);
